feat: portfolio page add trading account balance and positions table
