+++ components/partials/common/derivatives/position-table.header.vue
@@ -9,9 +9,6 @@
           {{ $t('trade.side') }}
         </div>
       </th>
-      <th class="text-right">
-        <span>{{ $t('trade.entry_price') }}</span>
-      </th>
       <th>
         <div class="flex items-center">
           <span>{{ $t('trade.amount') }}</span>
@@ -21,9 +18,12 @@
           />
         </div>
       </th>
+      <th class="text-right">
+        <span>{{ $t('position.entryMark') }}</span>
+      </th>
       <th>
         <div class="flex items-center justify-end">
-          <span>{{ $t('trade.liquidation_price') }}</span>
+          <span>{{ $t('position.estLiqPrice') }}</span>
           <v-icon-info-tooltip
             class="ml-2"
             :tooltip="$t('trade.liquidation_price_tooltip')"
