+++ components/partials/common/derivatives/position.vue
@@ -29,24 +29,33 @@
     </td>
 
     <td class="text-right font-mono">
+      <span v-if="hideBalance">{{ HIDDEN_BALANCE_DISPLAY }}</span>
       <v-number
-        :decimals="
-          market ? market.priceDecimals : UI_DEFAULT_PRICE_DISPLAY_DECIMALS
-        "
-        :number="price"
-      />
-    </td>
-
-    <td class="text-right font-mono">
-      <v-number
+        v-else
         :decimals="
           market ? market.quantityDecimals : UI_DEFAULT_AMOUNT_DISPLAY_DECIMALS
         "
         :number="quantity"
       />
     </td>
+
     <td class="text-right font-mono">
+      <span v-if="hideBalance">{{ HIDDEN_BALANCE_DISPLAY }}</span>
+      <div v-else>
+        <v-number
+          :decimals="
+            market ? market.priceDecimals : UI_DEFAULT_PRICE_DISPLAY_DECIMALS
+          "
+          :number="price"
+        />
+        <span class="text-gray-500 text-xs">{{ markPriceToFormat }}</span>
+      </div>
+    </td>
+
+    <td class="text-right font-mono">
+      <span v-if="hideBalance">{{ HIDDEN_BALANCE_DISPLAY }}</span>
       <v-number
+        v-else
         :decimals="
           market ? market.priceDecimals : UI_DEFAULT_PRICE_DISPLAY_DECIMALS
         "
@@ -54,14 +63,17 @@
       />
     </td>
     <td class="text-right">
+      <span v-if="hideBalance" class="font-mono">{{
+        HIDDEN_BALANCE_DISPLAY
+      }}</span>
       <div
-        v-if="!pnl.isNaN()"
+        v-else-if="!pnl.isNaN()"
         class="flex items-center justify-end text-xs"
         :class="pnlClass"
       >
-        <span class="mr-1">≈</span>
-        <div class="flex items-center">
-          <span class="mr-1 flex items-center">
+        <div class="flex items-end flex-col">
+          <span class="flex items-center">
+            <span class="mr-1">≈</span>
             <span>{{ pnl.gte(0) ? '+' : '' }}</span>
             <span
               :class="{
@@ -71,12 +83,10 @@
             >
               {{ pnlToFormat }}
             </span>
-            <span class="ml-1">{{ market.quoteToken.symbol }}</span>
+            <span class="ml-1 text-2xs">{{ market.quoteToken.symbol }}</span>
           </span>
-          <span class="flex" style="margin-top: -2px">
-            ({{
-              (percentagePnl.gte(0) ? '+' : '') + percentagePnl.toFormat(2)
-            }}%)
+          <span class="flex">
+            {{ (percentagePnl.gte(0) ? '+' : '') + percentagePnl.toFormat(2) }}%
           </span>
         </div>
       </div>
@@ -85,7 +95,9 @@
       }}</span>
     </td>
     <td class="text-right font-mono">
+      <span v-if="hideBalance">{{ HIDDEN_BALANCE_DISPLAY }}</span>
       <v-number
+        v-else
         :decimals="
           market ? market.priceDecimals : UI_DEFAULT_PRICE_DISPLAY_DECIMALS
         "
@@ -97,7 +109,10 @@
       </v-number>
     </td>
     <td class="text-right">
-      <div class="flex items-center justify-end h-8">
+      <span v-if="hideBalance" class="font-mono">
+        {{ HIDDEN_BALANCE_DISPLAY }}
+      </span>
+      <div v-else class="flex items-center justify-end h-8">
         <v-number
           :decimals="
             market ? market.priceDecimals : UI_DEFAULT_PRICE_DISPLAY_DECIMALS
@@ -107,7 +122,7 @@
         <button
           role="button"
           type="button"
-          class="border border-primary-500 text-primary-500 hover:text-primary-300 ml-2 px-1"
+          class="border border-gray-500 text-gray-500 hover:text-primary-500 hover:border-primary-500 ml-2 px-1"
           @click.stop.prevent="onAddMarginButtonClick"
         >
           &plus;
@@ -115,8 +130,9 @@
       </div>
     </td>
     <td class="text-right font-mono">
+      <span v-if="hideBalance">{{ HIDDEN_BALANCE_DISPLAY }}</span>
       <span
-        v-if="effectiveLeverage.gte(0)"
+        v-else-if="effectiveLeverage.gte(0)"
         class="flex items-center justify-end"
       >
         {{ effectiveLeverage.toFormat(2) }}
@@ -128,15 +144,13 @@
     </td>
 
     <td class="text-center relative">
-      <v-button :status="status" @click="onClosePositionClick">
+      <v-button
+        v-if="!hideBalance"
+        :status="status"
+        @click="onClosePositionClick"
+      >
         <div
-          class="flex items-center justify-center rounded-full bg-opacity-10 w-8 h-8 hover:bg-opacity-10"
-          :class="{
-            'bg-aqua-500 text-aqua-500 hover:bg-aqua-600 hover:text-aqua-600':
-              position.direction === TradeDirection.Long,
-            'bg-red-550 hover:bg-red-600 text-red-550 hover:text-red-600':
-              position.direction === TradeDirection.Short
-          }"
+          class="flex items-center justify-center rounded-full bg-opacity-10 w-8 h-8 hover:bg-opacity-10 bg-red-550 hover:bg-red-600 text-red-550 hover:text-red-600"
         >
           <v-icon-close class="h-4 w-4" />
         </div>
@@ -152,13 +166,12 @@ import { TradeDirection } from '@injectivelabs/ts-types'
 import {
   UiDerivativeLimitOrder,
   UiDerivativeMarketWithToken,
-  UiDerivativeOrderbook,
   UiPosition,
   UiSpotLimitOrder,
-  ZERO_IN_BASE,
-  UiPriceLevel
+  ZERO_IN_BASE
 } from '@injectivelabs/ui-common'
 import {
+  HIDDEN_BALANCE_DISPLAY,
   UI_DEFAULT_AMOUNT_DISPLAY_DECIMALS,
   UI_DEFAULT_PRICE_DISPLAY_DECIMALS
 } from '~/app/utils/constants'
@@ -168,11 +181,17 @@ export default Vue.extend({
     position: {
       required: true,
       type: Object as PropType<UiPosition>
+    },
+
+    hideBalance: {
+      type: Boolean,
+      default: false
     }
   },
 
   data() {
     return {
+      HIDDEN_BALANCE_DISPLAY,
       UI_DEFAULT_PRICE_DISPLAY_DECIMALS,
       UI_DEFAULT_AMOUNT_DISPLAY_DECIMALS,
       TradeDirection,
@@ -185,16 +204,8 @@ export default Vue.extend({
       return this.$accessor.derivatives.market
     },
 
-    currentOrderbook(): UiDerivativeOrderbook | undefined {
-      return this.$accessor.derivatives.orderbook
-    },
-
     currentOrders(): UiDerivativeLimitOrder[] {
       return this.$accessor.derivatives.subaccountOrders
-    },
-
-    orderbooks(): Record<string, UiDerivativeOrderbook> {
-      return this.$accessor.positions.orderbooks
     },
 
     markets(): UiDerivativeMarketWithToken[] {
@@ -221,21 +232,6 @@ export default Vue.extend({
       const { markets, position } = this
 
       return markets.find((m) => m.marketId === position.marketId)
-    },
-
-    orderbook(): UiDerivativeOrderbook | undefined {
-      const { isOnMarketPage, currentMarket, currentOrderbook } = this
-      const { orderbooks, position } = this
-
-      if (
-        isOnMarketPage &&
-        currentMarket &&
-        currentMarket.marketId === position.marketId
-      ) {
-        return currentOrderbook
-      }
-
-      return orderbooks[position.marketId]
     },
 
     orders(): Array<UiDerivativeLimitOrder | UiSpotLimitOrder> {
@@ -270,26 +266,6 @@ export default Vue.extend({
       return new BigNumberInWei(position.margin).toBase(
         market.quoteToken.decimals
       )
-    },
-
-    buys(): UiPriceLevel[] {
-      const { orderbook } = this
-
-      if (!orderbook) {
-        return []
-      }
-
-      return orderbook.buys
-    },
-
-    sells(): UiPriceLevel[] {
-      const { orderbook } = this
-
-      if (!orderbook) {
-        return []
-      }
-
-      return orderbook.sells
     },
 
     quantity(): BigNumberInBase {
@@ -393,26 +369,14 @@ export default Vue.extend({
     },
 
     pnl(): BigNumberInBase {
-      const { market, sells, position, price, buys } = this
+      const { markPrice, market, position, price } = this
 
       if (!market) {
         return ZERO_IN_BASE
       }
 
-      const [sell] = sells
-      const [buy] = buys
-      const highestBuy = new BigNumberInBase(buy ? buy.price : 0)
-      const lowestSell = new BigNumberInBase(sell ? sell.price : 0)
-      const executionPrice = new BigNumberInWei(
-        highestBuy.plus(lowestSell).div(2)
-      ).toBase(market.quoteToken.decimals)
-
-      if (executionPrice.isZero()) {
-        return new BigNumberInBase('')
-      }
-
       return new BigNumberInBase(position.quantity)
-        .times(executionPrice.minus(price))
+        .times(markPrice.minus(price))
         .times(position.direction === TradeDirection.Long ? 1 : -1)
     },
 
@@ -424,6 +388,16 @@ export default Vue.extend({
       }
 
       return pnl.toFormat(market.priceDecimals)
+    },
+
+    markPriceToFormat(): string {
+      const { market, markPrice } = this
+
+      if (!market) {
+        return markPrice.toFormat(UI_DEFAULT_PRICE_DISPLAY_DECIMALS)
+      }
+
+      return markPrice.toFormat(market.priceDecimals)
     },
 
     percentagePnl(): BigNumberInBase {
